fix: switch ints to longs in line with db

+++ backend/src/main/java/com/davidhorobin/budgetbalance/controller/TransactionController.java
@@ -25,7 +25,7 @@ public class TransactionController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<TransactionResponse> getTransactionById(@PathVariable Integer id) {
+    public ResponseEntity<TransactionResponse> getTransactionById(@PathVariable long id) {
         return ResponseEntity.ok(transactionService.getTransactionById(id));
     }
 
@@ -35,7 +35,7 @@ public class TransactionController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteTransactionById(@PathVariable Integer id) {
+    public ResponseEntity<String> deleteTransactionById(@PathVariable long id) {
         transactionService.deleteTransactionById(id);
         return ResponseEntity.ok("Deleted transaction successfully");
     }
