+++ backend/src/main/java/com/davidhorobin/budgetbalance/repository/TransactionRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TransactionRepo extends JpaRepository<Transaction, Integer> {
+public interface TransactionRepo extends JpaRepository<Transaction, Long> {
     List<Transaction> findAllByOrderByTimeAsc();
 }
